test(agent): consolidate duplicate provider, config, and api test suites
